pyproject.toml
[project]
name = "fastdjango"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "a2wsgi>=1.10.10",
    "annotated-types>=0.7.0",
    "boto3>=1.42.24",
    "cachetools>=6.2.4",
    "celery>=5.6.2",
    "celery-types>=0.24.0",
    "colorlog>=6.10.1",
    "dj-database-url>=3.1.0",
    "django>6",
    "django-storages[s3]>=1.14.6",
    "django-stubs[compatible-mypy]>=5.2.8",
    "fastapi>=0.128.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "logfire[fastapi, django, celery, requests, psycopg, httpx, redis]>=4.18.0",
    "psycopg[binary]>=3.3.2",
    "punq>=0.7.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "throttled-py>=3.1.1",
    "uvicorn[standard]>=0.34.0",
    "watchdog[watchmedo]>=6.0.0",
]

[dependency-groups]
dev = [
    "icecream>=2.1.9",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "prek>=0.2.28",
    "pyrefly>=0.47.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
    "ty>=0.0.9",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.0",
]

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
namespace = true

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
local_partial_types = true
warn_unreachable = true
enable_error_code = [
    "truthy-bool",
    "truthy-iterable",
    "redundant-expr",
    "unused-awaitable",
    "ignore-without-code",
    "possibly-undefined",
    "redundant-self",
    "unimported-reveal",
    "deprecated",
    "exhaustive-match",
]
disable_error_code = [
    "unused-ignore",
    "no-any-return",
]

[[tool.mypy.overrides]]
module = "api.*.migrations.*"
disable_error_code = ["no-untyped-def"]

[tool.django-stubs]
django_settings_module = "configs.django"

[tool.ty.rules]
unused-ignore-comment = "ignore"

[tool.logfire]
ignore_no_config = true

[tool.pytest.ini_options]
minversion = "9.0"
addopts = "--reuse-db --create-db --exitfirst -vv --cov=src --cov-report=html --cov-fail-under=80"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    # Entry points
    "src/manage.py",
    "src/delivery/http/app.py",
    "src/delivery/tasks/app.py",
    # Django admin (requires manual testing)
    "src/**/admin.py",
    # Observability/telemetry (requires external services)
    "src/**/logfire/**",
    # Celery event handlers (signal-based, hard to test in isolation)
    "src/delivery/tasks/events.py",
]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "@abstractmethod",
    "if TYPE_CHECKING:",
    "@overload",
    "class .*\\bProtocol\\):",
    "^\\s*pass$",
    "^\\s*\\.\\.\\.$",
]
skip_empty = true
